change import, move composer_name_to_vec
