+++ feature_utils.py
@@ -12,7 +12,7 @@ get feature information to generate or modify feature
 """
 
 import math
-import utils
+from . import utils
 
 def cal_beat_importance(beat_position, numerator):
     """ Returns beat importance in integer
@@ -308,3 +308,17 @@ def pedal_sigmoid(pedal_value, k=8):
 
 def get_trill_parameters():
     return
+
+
+def composer_name_to_vec(composer_name):
+    composer_name_list = ['Bach','Balakirev', 'Beethoven', 'Brahms', 'Chopin', 'Debussy', 'Glinka', 'Haydn',
+                          'Liszt', 'Mozart', 'Prokofiev', 'Rachmaninoff', 'Ravel', 'Schubert', 'Schumann', 'Scriabin']
+    one_hot_vec = [0] * (len(composer_name_list)  + 1)
+    if composer_name in composer_name_list:
+        index = composer_name_list.index(composer_name)
+    else:
+        index = len(composer_name_list)
+        print('The given composer name {} is not in the list'.format(composer_name))
+    one_hot_vec[index] = 1
+
+    return one_hot_vec
